pyproject.toml
[tool.poetry]
name = "persona-link"
version = "0.1.0"
description = "A comprehensive toolkit for easy integration of avatars or digital humans from any providers onto any AI base."
authors = ["Abhinav Dayal <abhinav@enligence.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
python-dotenv = "^1.0.1"
pydantic = "^2.7.0"
azure-storage-blob = { version = "^12.19.1", optional = true }
aiohttp = "^3.9.5"
aiofiles = { version = "^23.2.1", optional = true }
tortoise-orm = { version = "^0.20.0", optional = true }
asyncpg = { version = "^0.29.0", optional = true }
azure-cognitiveservices-speech = { version = "^1.37.0", optional = true }

[tool.poetry.extras]
azure = ["azure-storage-blob"]
postgres = ["tortoise-orm", "asyncpg"]
sqlite = ["tortoise-orm"]
azure-tts = ["azure-cognitiveservices-speech"]


[tool.poetry.group.dev.dependencies]
pyright = "^1.1.359"
pre-commit = "^3.7.0"
mkdocs = "^1.6.0"
mkdocstrings = "^0.24.3"
mkdocstrings-python = "^1.10.0"
isort = "^5.13.2"
autoflake = "^2.3.1"


[tool.poetry.group.test.dependencies]
pytest-asyncio = "^0.23.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
